// webapp/src/app/core/api/v1/model/newApplication.ts
/**
 * permaAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface NewApplication {
  applicationTypeId?: number;
  memberId?: number;
  applicationName?: string | null;
  coverageYear?: number;
  applicationStatus?: string | null;
  dueDate?: string | null;
  completedDate?: string | null;
  completedBy?: string | null;
  periodMonth?: number | null;
  periodQuarter?: number | null;
}